# Evolutionary strategies, compact Trie