--- a/boggle/gcs.py
+++ b/boggle/gcs.py
@@ -1,3 +1,5 @@
+import os
+
 from google.cloud import storage