… .claude/settings.local.json
… boggle/boggler_test.py
… boggle/breaker_test.py
… boggle/ibuckets_test.py
… boggle/orderly_tree_test.py
… boggle/trie.py
… boggle/trie_test.py
… cpp/arena.h
… cpp/board_class_boggler.h
… cpp/boggler.h
… cpp/cpp_boggle.cc
… cpp/ibuckets.h
… cpp/orderly_tree_builder.h
… cpp/trie.cc
… cpp/trie.h
… wordlists/twl06.14char.txt
… wordlists/twl06.20char.txt