Closes #118 
Closes #142 

The layout for the Trie is now:

```
uint32_t child_indices_;  // 1 bit for is_word, 31 bits for children
uint16_t mark_;
uint16_t children_;  // offset to first child / sizeof(Trie)
```

So it's 8 bytes per node. A few awkward things here:

- The `children_` offset field is too small for YAWL and SOWPODS.
- The 16-bit `mark_` requires resetting all marks every 65535 board evaluations.

Resetting marks takes ~2ms on my Macbook and it takes several seconds to evaluate that many 5x5 boards, so this is a pretty small penalty.

Total RAM goes from ~100MB for `IndexedTrie` → 2.97MB. Performance is 

- ~300k variations on "ligdrmanesietildsracsepes": 48.69→44.77s, 7932→8626 bds/sec (8% speedup)
- 1M random boards: 10.31→8.58s, 96959→116565 bds/sec (17% speedup)

TBD perf on GCP